requirements.txt


# Backend and Web Framework
Flask
gunicorn              # Production WSGI Server for deployment

# Data Science and Machine Learning Core Libraries
pandas
numpy
scikit-learn
joblib                # For saving and loading ML models

# Data Collection and API Handling
requests              # General HTTP requests (for MAL API)

# Social Data Collection and NLP
praw                  # Python Reddit API Wrapper

# Sentiment Analysis and NLP
# nltk                  # NLTK base library (No longer primary, but keep for now if needed later)
# vaderSentiment        # VADER (Replaced by TextBlob due to HTTP 429 error)
textblob              # Alternative: Simple API for sentiment analysis

# Utilities and Security
python-dotenv         # To load configuration variables from the .env file

# Data Visualization (For generating trend graphs in the backend)
matplotlib
seaborn


scikit-surprise          # For building and analyzing recommender systems